+++ CHANGELOG.md
@@ -64,3 +64,7 @@
 - Fixed a bug when use batched address. Since we were sorting the address during the forward pass, the output prediction 
   tags were not aligned with the supposed parsed address. We have removed the sorting, and now the results are more 
   aligned with our research.
+
+## dev
+
+- Added verbose flag to training and test base on the __init__ of address parser.